Chore: Move FileType File
