+++ data/file/src/main/java/com/seugi/data/file/repository/FileRepositoryImpl.kt
@@ -5,8 +5,8 @@ import com.seugi.common.model.asResult
 import com.seugi.common.utiles.DispatcherType
 import com.seugi.common.utiles.SeugiDispatcher
 import com.seugi.data.file.FileRepository
+import com.seugi.data.file.model.FileType
 import com.seugi.file.FileDataSource
-import com.seugi.file.request.FileType
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
@@ -18,7 +18,7 @@ class FileRepositoryImpl @Inject constructor(
     private val fileDataSource: FileDataSource,
 ) : FileRepository {
     override suspend fun fileUpload(file: String, type: FileType): Flow<Result<String>> = flow {
-        val response = fileDataSource.fileUpload(file = file, type = type).data
+        val response = fileDataSource.fileUpload(file = file, type = type.name).data
 
         emit(response)
     }
