+++ network/file/src/main/java/com/seugi/file/FileDataSource.kt
@@ -1,8 +1,7 @@
 package com.seugi.file
 
-import com.seugi.file.request.FileType
 import com.seugi.network.core.response.BaseResponse
 
 interface FileDataSource {
-    suspend fun fileUpload(type: FileType, file: String): BaseResponse<String>
+    suspend fun fileUpload(type: String, file: String): BaseResponse<String>
 }
