+++ network/file/src/main/java/com/seugi/file/datasource/FileDataSourceImpl.kt
@@ -1,7 +1,6 @@
 package com.seugi.file.datasource
 
 import com.seugi.file.FileDataSource
-import com.seugi.file.request.FileType
 import com.seugi.network.core.SeugiUrl
 import com.seugi.network.core.response.BaseResponse
 import io.ktor.client.HttpClient
@@ -11,8 +10,6 @@ import io.ktor.client.request.forms.MultiPartFormDataContent
 import io.ktor.client.request.forms.formData
 import io.ktor.client.request.post
 import io.ktor.client.request.setBody
-import io.ktor.http.ContentDisposition.Companion.File
-import io.ktor.http.ContentType
 import io.ktor.http.Headers
 import io.ktor.http.HttpHeaders
 import java.io.File
@@ -21,7 +18,7 @@ import javax.inject.Inject
 class FileDataSourceImpl @Inject constructor(
     private val httpClient: HttpClient,
 ) : FileDataSource {
-    override suspend fun fileUpload(type: FileType, file: String): BaseResponse<String> = httpClient.post(urlString = "${SeugiUrl.File.FILE_UPLOAD}/$type") {
+    override suspend fun fileUpload(type: String, file: String): BaseResponse<String> = httpClient.post(urlString = "${SeugiUrl.File.FILE_UPLOAD}/$type") {
         setBody(
             MultiPartFormDataContent(
                 formData {
